requirements.txt
absl-py==1.2.0
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
arrow==1.2.3
asttokens==2.0.8
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
cachetools==5.2.0
certifi==2022.9.14
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cookiecutter==1.7.3
cryptography==38.0.1
cssselect==1.1.0
cycler==0.11.0
databases==0.6.1
decorator==5.1.1
dnspython==2.2.1
easyocr==1.6.2
ecdsa==0.18.0
email-validator==1.3.0
executing==1.1.0
fastapi==0.85.0
fonttools==4.37.3
frozenlist==1.3.1
google-api-core==2.10.2
google-auth==2.12.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-language==2.6.1
google-cloud-speech==2.16.2
google-cloud-translate==3.8.4
google-cloud-vision==3.1.4
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
h11==0.13.0
idna==3.4
imageio==2.22.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
kiwisolver==1.4.4
limits==1.6
lxml==4.9.1
manage-fastapi==1.1.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
multidict==6.0.2
networkx==2.8.6
ninja==1.10.2.3
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
ormar==0.11.3
packaging==21.3
pandas==1.5.0
parsel==1.6.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
poyo==0.5.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==4.21.7
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pyclipper==1.3.0.post3
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==6.1.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
questionary==1.10.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
seaborn==0.12.0
selectorlib==0.16.0
Shapely==1.8.4
six==1.16.0
slowapi==0.1.6
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
stack-data==0.5.1
starlette==0.20.4
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
thop==0.1.1.post2209072238
tifffile==2022.8.12
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
traitlets==5.4.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
w3lib==2.0.1
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
